agent_backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
pymongo
passlib
python-jose
bcrypt
requests
langchain
langchain-community
ollama
openai
langchain-openai
